Move compression configuration to new class as discussed with @jackshirazi
